Refactor help and privacy pages, update routes
Moved help and privacy components to new locations under _components, updated imports and file names to use English routes (/help, /privacy-policy), and removed the old Spanish route files. Adjusted Footer links to match new routes. Improved formatting and icons in privacy content.
